feat: refactor embedding in post and blog into seperate services
