+++ src/blog/blog.module.ts
@@ -4,10 +4,11 @@ import { AuthModule } from 'src/auth/auth.module';
 import { EmbeddingModule } from 'src/embedding/embedding.module';
 import { BlogManagementService } from './blog-management.service';
 import { BlogExploreService } from './blog-explore.service';
+import { BlogEmbeddingService } from './blog-embedding.service';
 
 @Module({
   imports: [AuthModule, EmbeddingModule],
-  providers: [BlogManagementService, BlogExploreService],
+  providers: [BlogManagementService, BlogExploreService, BlogEmbeddingService],
   controllers: [BlogController],
 })
 export class BlogModule {}
