+++ src/posts/posts-management.service.ts
@@ -6,28 +6,15 @@ import {
 } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import { plainToInstance } from 'class-transformer';
-import { StorageService } from 'src/storage/storage.service';
-import { EmbeddingService } from 'src/embedding/embedding.service';
-import { MediaType, Post } from '@prisma/client';
-import { QdrantClient } from '@qdrant/js-client-rest';
+import { StorageService } from 'src/storage/storage.service';import { MediaType, Post } from '@prisma/client';
 import { TryCatch } from 'src/common/try-catch.decorator';
 import { PostDetailsResponseDto } from './dto/response/post-details.dto';
 import { UpdatePostDto } from './dto/request/update-post.dto';
 import { FileUploadResponse } from 'src/storage/dto/response.dto';
-import axios from 'axios';
 import { PatchThumbnailDto } from './dto/request/patch-thumbnail.dto';
-import { nanoid } from 'nanoid';
-import { Readable } from 'stream';
-import { VECTOR_DIMENSION } from 'src/embedding/embedding.utils';
-import { SyncEmbeddingResponseDto } from 'src/common/response/sync-embedding.dto';
 import { QdrantService } from 'src/embedding/qdrant.service';
 import { CreatePostRequestDto } from './dto/request/create-post.dto';
-
-class VectorParams {
-  titleEmbedding: number[];
-  descriptionEmbedding: number[];
-  imagesEmbedding: number[];
-}
+import { PostsEmbeddingService } from './posts-embedding.service';
 
 class MediaTocreate {
   url: string;
@@ -47,31 +34,9 @@ export class PostsManagementService {
   constructor(
     private readonly prisma: PrismaService,
     private readonly storageService: StorageService,
-    private readonly embeddingService: EmbeddingService,
-    private readonly qdrantClient: QdrantClient,
     private readonly qdrantService: QdrantService,
+    private readonly postEmbeddingService: PostsEmbeddingService,
   ) {}
-
-  @TryCatch("something went wrong with ensuring post collection exists")
-  private async ensurePostCollectionExists() {
-    const collectionExists = this.qdrantService.collectionExists(
-      this.qdrantCollectionName,
-    );
-
-    if (!collectionExists) {
-      await this.qdrantClient.createCollection(this.qdrantCollectionName, {
-        vectors: {
-          title: { size: VECTOR_DIMENSION, distance: 'Cosine' },
-          description: { size: VECTOR_DIMENSION, distance: 'Cosine' },
-          images: { size: VECTOR_DIMENSION, distance: 'Dot' },
-        },
-      });
-
-      console.log(
-        `Created Qdrant collection '${this.qdrantCollectionName}' with named vectors`,
-      );
-    }
-  }
 
   @TryCatch()
   async createPost(
@@ -103,7 +68,7 @@ export class PostsManagementService {
       include: { medias: true, user: true, categories: true },
     });
 
-    void this.upsertPostEmbedding(
+    void this.postEmbeddingService.upsertPostEmbedding(
       createdPost.id,
       createdPost.title,
       createdPost.description ?? undefined,
@@ -296,7 +261,7 @@ export class PostsManagementService {
       include: { medias: true, user: true, categories: true },
     });
 
-    this.upsertPostEmbedding(
+    this.postEmbeddingService.upsertPostEmbedding(
       updatedPost.id,
       updatedPost.title,
       updatedPost.description ?? undefined,
@@ -335,167 +300,6 @@ export class PostsManagementService {
       await this.storageService.deleteFiles(urls);
     }
     await this.qdrantService.deletePoints(this.qdrantCollectionName, [postId]);
-  }
-
-  private async getVectorParams(
-    title: string,
-    description: string | undefined,
-    imageFiles: Express.Multer.File[],
-  ): Promise<VectorParams> {
-    const [titleEmbedding, descriptionEmbedding, imagesEmbedding]: number[][] =
-      await Promise.all([
-        this.embeddingService.generateEmbeddingFromText(title),
-
-        description
-          ? this.embeddingService.generateEmbeddingFromText(description)
-          : Promise.resolve(new Array(VECTOR_DIMENSION).fill(0)),
-
-        imageFiles && imageFiles.length > 0
-          ? Promise.all(
-              imageFiles.map((image) =>
-                this.embeddingService.generateEmbeddingFromImageBlob(
-                  new Blob([image.buffer]),
-                ),
-              ),
-            ).then((embeds: number[][]) => this.averageEmbeddings(embeds))
-          : Promise.resolve(new Array(VECTOR_DIMENSION).fill(0)),
-      ]);
-
-    return {
-      titleEmbedding: titleEmbedding,
-      descriptionEmbedding: descriptionEmbedding,
-      imagesEmbedding: imagesEmbedding,
-    };
-  }
-
-  private averageEmbeddings(embeddings: number[][]): number[] {
-    if (!embeddings || embeddings.length === 0) return []; // handle empty case safely
-    const length = embeddings[0].length;
-    const sum = new Array(length).fill(0);
-    embeddings.forEach((vec) => {
-      for (let i = 0; i < length; i++) {
-        sum[i] += vec[i];
-      }
-    });
-    return sum.map((val) => val / embeddings.length);
-  }
-
-  @TryCatch()
-  private async upsertPostEmbedding(
-    postId: number,
-    title: string,
-    description: string | undefined,
-    imageFiles: Express.Multer.File[],
-  ): Promise<void> {
-    await this.ensurePostCollectionExists();
-    const { titleEmbedding, descriptionEmbedding, imagesEmbedding } =
-      await this.getVectorParams(title, description, imageFiles);
-
-    const operationInfo = await this.qdrantClient.upsert(
-      this.qdrantCollectionName,
-      {
-        wait: true,
-        points: [
-          {
-            id: postId,
-            vector: {
-              title: titleEmbedding,
-              description: descriptionEmbedding,
-              images: imagesEmbedding,
-            } as Record<string, number[]>,
-          },
-        ],
-      },
-    );
-
-    console.log('Upsert operation info:', operationInfo);
-  }
-
-  @TryCatch()
-  async syncPostEmbeddings(): Promise<SyncEmbeddingResponseDto> {
-    await this.qdrantService.deleteAllPoints(this.qdrantCollectionName);
-  
-    const posts = await this.prisma.post.findMany({
-      include: { medias: true },
-    });
-
-    if (!posts || posts.length === 0) {
-      console.log('No posts found.');
-      return {
-        message: 'No posts found to sync',
-        count: 0,
-        syncedItems: [],
-      };
-    }
-    console.log(`Found ${posts.length} posts to sync.`);
-
-    const points = await Promise.all(
-      posts.map(async (post) => {
-        const imageMedias = post.medias.filter(
-          (m) => m.media_type === MediaType.image,
-        );
-        const imageFiles: Express.Multer.File[] =
-          await this.buildImageFilesFromUrls(imageMedias.map((m) => m.url));
-
-        // get your embeddings
-        const { titleEmbedding, descriptionEmbedding, imagesEmbedding } =
-          await this.getVectorParams(
-            post.title,
-            post.description ?? undefined,
-            imageFiles,
-          );
-
-        return {
-          id: post.id,
-          vector: {
-            title: titleEmbedding,
-            description: descriptionEmbedding,
-            images: imagesEmbedding,
-          } as Record<string, number[]>,
-        };
-      }),
-    );
-
-    const operationInfo = await this.qdrantClient.upsert(
-      this.qdrantCollectionName,
-      {
-        wait: true,
-        points: points,
-      },
-    );
-
-    console.log('Upsert result:', operationInfo);
-    return {
-      message: 'Post embeddings synced successfully',
-      count: points.length,
-      syncedItems: points.map((point) => point.id.toString()),
-    };
-  }
-
-  private async buildImageFilesFromUrls(
-    imageUrls: string[],
-  ): Promise<Express.Multer.File[]> {
-    return await Promise.all(
-      imageUrls.map(async (url) => {
-        const res = await axios.get<ArrayBuffer>(url, {
-          responseType: 'arraybuffer',
-        });
-        const buffer = Buffer.from(res.data);
-        const ext = url.split('.').pop() ?? 'png';
-        return {
-          fieldname: 'file',
-          originalname: `${nanoid()}.${ext}`,
-          encoding: '7bit',
-          mimetype: `image/${ext}`,
-          buffer,
-          size: buffer.length,
-          destination: '',
-          filename: '',
-          path: '',
-          stream: Readable.from(buffer),
-        } as Express.Multer.File;
-      }),
-    );
   }
 
   async updateThumbnailCropMeta(
