+++ src/posts/posts.module.ts
@@ -7,6 +7,7 @@ import { AuthModule } from 'src/auth/auth.module';
 import { EmbeddingModule } from 'src/embedding/embedding.module';
 import { PostsExploreService } from './posts-explore.service';
 import { WorkflowAssistService } from './workflow-assist.service';
+import { PostsEmbeddingService } from './posts-embedding.service';
 
 @Module({
   imports: [AuthModule, EmbeddingModule],
@@ -16,6 +17,7 @@ import { WorkflowAssistService } from './workflow-assist.service';
     StorageService,
     S3StorageProvider,
     WorkflowAssistService,
+    PostsEmbeddingService,
   ],
   controllers: [PostsController],
 })
